refactor: refactor test
